Refactor socket acquisition into a GetOrCreateOpenSocket method that a LoggerAPIConnectionWS instance can use to send metrics to the api.
GetOrCreateOpenSocket will:
1. create an web socket instance if one does not exist
2. open the socket if not already opened
3. time-out the open operation after 2500ms (default)

Write both unit and functional tests that verify:
1. web socket is created reasonably
2. web socket opens in a reasonable amount of time
3. metrics can be sent over the socket and received in a reasonable amount of time.

Note that the metric send/receive test is basically a functional test, though it is self-hosting in that a local web socket server is spun-up to support the test.  the test does not rely on the api.

Much of the approach for this test was inspired by WebSocket4Net’s WebSocketClientTest test class.